refactor(pipeline): Frontend virtual base class, Pipeline owns initialization, int dimensions
Major architectural refactoring across 6 phases:

Phase 1: CMake platform detection — replace `if(TARGET ...)` with
`if("rockchip" IN_LIST CMAKE_PLATFORM)` for explicit platform selection.

Phase 2: Unify all width/height types from `size_t` to `int` across the
entire pipeline (InferenceEngine, Frontend, Preprocessor, RVMMode,
MODNetMode, MattingServer). Industry standard: OpenCV/FFmpeg/MPPKit/RGAKit
all use `int` for image dimensions.

Phase 3: Convert Frontend to pure virtual base class (FrontendBase) with
platform-specific subclasses (RockchipFrontend, NoHwFrontend). Eliminate
the factory layer (05-factory/ directory). Use build-time source file
selection pattern matching InferenceEngine. FrontendBase::Create() static
factory method replaces the free function CreateFrontend().

Phase 4: Pipeline::Init() now owns all initialization — engine load,
backend configuration, and dependency injection to modes via setters.
RVMMode::Run() and MODNetMode::Run() become parameter-free executors.
MattingBackend owned by Pipeline, accessed by modes via non-owning pointer.

Phase 6: Create AI codebase documentation in docs/zh/1-for-ai/ (7 new
files: architecture, frontend, build-system, data-flow, timing,
naming-convention, README). Update CLAUDE.md to point to docs. Clean up
docs directory: remove en/ mirror, empty beginner dir, merge highlights
into architecture.md.

+++ runtime/cpp/apps/matting/server/include/pipeline/stages/frontend/04-preprocess/preprocessor.h
@@ -39,14 +39,14 @@ public:
 
     // Preprocess a BGR frame into a TensorData for inference.
     TensorData preprocess(const cv::Mat& bgr_frame,
-                          size_t model_width,
-                          size_t model_height) override;
+                          int model_width,
+                          int model_height) override;
 
     // Configure output binary dump directory.
     void SetOutputBinPath(const std::string& path);
 
 private:
-    TensorData PreprocessCore(cv::Mat img, size_t model_width, size_t model_height);
+    TensorData PreprocessCore(cv::Mat img, int model_width, int model_height);
 
     bool IsDumpEnabled() const { return !output_bin_path_.empty(); }
